requirements.txt
discord.py
pyternos
